import subversion r1146
svn checkout svn://svn.code.sf.net/p/dansguardian/code/trunk dansguardian-code

+++ dansguardian/src/UDSocket.cpp
@@ -1,21 +1,8 @@
 // UDSocket class - implements BaseSocket for UNIX domain sockets
 
-//Please refer to http://dansguardian.org/?page=copyright2
-//for the license for this code.
-
-//  This program is free software; you can redistribute it and/or modify
-//  it under the terms of the GNU General Public License as published by
-//  the Free Software Foundation; either version 2 of the License, or
-//  (at your option) any later version.
-//
-//  This program is distributed in the hope that it will be useful,
-//  but WITHOUT ANY WARRANTY; without even the implied warranty of
-//  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-//  GNU General Public License for more details.
-//
-//  You should have received a copy of the GNU General Public License
-//  along with this program; if not, write to the Free Software
-//  Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+// For all support, instructions and copyright go to:
+// http://dansguardian.org/
+// Released under the GPL v2, with the OpenSSL exception described in the README file.
 
 
 // INCLUDES
@@ -25,6 +12,7 @@
 #endif
 #include "UDSocket.hpp"
 
+#include <string.h>
 #include <syslog.h>
 #include <csignal>
 #include <fcntl.h>
@@ -54,7 +42,10 @@ UDSocket::UDSocket()
 	memset(&my_adr, 0, sizeof my_adr);
 	memset(&peer_adr, 0, sizeof peer_adr);
 	my_adr.sun_family = AF_UNIX;
+	strcpy(my_adr.sun_path, "");
 	peer_adr.sun_family = AF_UNIX;
+	strcpy(peer_adr.sun_path, "");
+	my_adr_length = 0;
 }
 
 // create socket from pre-existing FD (address structs will be invalid!)
@@ -63,7 +54,10 @@ UDSocket::UDSocket(int fd):BaseSocket(fd)
 	memset(&my_adr, 0, sizeof my_adr);
 	memset(&peer_adr, 0, sizeof peer_adr);
 	my_adr.sun_family = AF_UNIX;
-	peer_adr.sun_family = AF_UNIX;
+        strcpy(my_adr.sun_path, "");
+        peer_adr.sun_family = AF_UNIX;
+        strcpy(peer_adr.sun_path, "");
+	my_adr_length = 0;
 }
 
 // create socket from given FD & local address (checkme: is it local or remote that gets passed in here?)
@@ -81,7 +75,10 @@ void UDSocket::reset()
 	memset(&my_adr, 0, sizeof my_adr);
 	memset(&peer_adr, 0, sizeof peer_adr);
 	my_adr.sun_family = AF_UNIX;
-	peer_adr.sun_family = AF_UNIX;
+        strcpy(my_adr.sun_path, "");
+        peer_adr.sun_family = AF_UNIX;
+        strcpy(peer_adr.sun_path, "");
+	my_adr_length = 0;
 }
 
 // accept incoming connection & return new UDSocket
@@ -96,6 +93,9 @@ UDSocket* UDSocket::accept()
 // connect to given server (following default constructor)
 int UDSocket::connect(const char *path)
 {
+	if(strlen(path) > 108)
+		return -1;
+
 #ifdef DGDEBUG
 	std::cout << "uds connect:" << path << std::endl;
 #endif
@@ -109,7 +109,11 @@ int UDSocket::connect(const char *path)
 // bind socket to given path
 int UDSocket::bind(const char *path)
 {				// to bind a unix domain socket to a path
+	if(strlen(path) > 108)
+		return -1;
+
 	unlink(path);
+
 	strcpy(my_adr.sun_path, path);
 
 	my_adr_length = offsetof(struct sockaddr_un, sun_path) + strlen(path);
